refactor: optimize login function
this changes optimize login function by separating the login form to another file to make it modular.
