+++ auth.jsx
@@ -1,7 +1,7 @@
-import axios from "axios";
-import jwt from "jsonwebtoken";
-import NextAuth from "next-auth";
-import Credentials from "next-auth/providers/credentials";
+import axios from "axios"
+import jwt from "jsonwebtoken"
+import NextAuth from "next-auth"
+import Credentials from "next-auth/providers/credentials"
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
   providers: [
@@ -12,34 +12,28 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
       },
       authorize: async (credentials) => {
         try {
-          // Send login request
-          const response = await axios.post(
-            `https://172.168.10.119:2000/api/log-in`,
-            {
-              username: credentials.username,
-              password: credentials.password,
-            }
-          );
-      
+          const response = await axios.post(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/log-in`, {
+            username: credentials.username,
+            password: credentials.password,
+          })
+
           if (response.data && response.data.token) {
-            const decodedToken = jwt.decode(response.data.token);
-      
+            const decodedToken = jwt.decode(response.data.token)
             return {
               id: decodedToken.id,
               first_name: decodedToken.first_name,
               last_name: decodedToken.last_name,
               role_id: decodedToken.role_id,
-              emp_no: decodedToken.emp_no,
+              exp: decodedToken.exp,
+              originalExp: decodedToken.exp, // Store the original expiration
               token: response.data.token,
-            };
+            }
           }
-      
-          // Explicitly throw an error if authentication fails
-          throw new Error("Invalid Employee Number or Password.");
+          throw new Error("Invalid Employee Number or Password.")
         } catch (error) {
           return null
         }
-      },      
+      },
     }),
   ],
   pages: {
@@ -49,16 +43,24 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
     strategy: "jwt",
   },
   callbacks: {
-    jwt: ({ token, user }) => {
+    jwt: ({ token, user, trigger }) => {
       if (user) {
-        token.id = user.id;
-        token.first_name = user.first_name;
-        token.last_name = user.last_name;
-        token.role_id = user.role_id;
-        token.emp_no = user.emp_no;
-        token.token = user.token;
+        return {
+          ...token,
+          id: user.id,
+          first_name: user.first_name,
+          last_name: user.last_name,
+          role_id: user.role_id,
+          exp: user.exp,
+          originalExp: user.exp,
+          token: user.token,
+        }
       }
-      return token;
+
+      return {
+        ...token,
+        exp: token.originalExp,
+      }
     },
     session: ({ session, token }) => {
       session.user = {
@@ -66,10 +68,11 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
         first_name: token.first_name,
         last_name: token.last_name,
         role_id: token.role_id,
-        emp_no: token.emp_no,
-      };
-      session.token = token.token;
-      return session;
+        exp: token.originalExp,
+      }
+      session.token = token.token
+      return session
     },
   },
-});
+})
+
